Retouches
this repo used to be a python only then I changed its name to a more general one as it will include more lessons as it goes. So, I created a python directory as a first step. Then I added my latest lecture codes regarding the functions topic. More will follow as the time passes.